Move styles to /libs/design-system/styles
