+++ libs/design-system/.storybook/preview.js
@@ -1,1 +1,1 @@
-import './tailwind.css';
+import '../styles/tailwind.css';
