+++ libs/design-system/src/lib/Styles/Colors/Colors.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, Story } from '@storybook/react';
-import { Colors } from './Colors';
+import { Colors } from './';
 
 export default {
   component: Colors,
