+++ libs/design-system/styles/fonts/ibm-plex/IBMPlexMono-Regular.ttf
