+++ libs/design-system/styles/fonts/termina/Termina-ExtraLight.otf
