+++ libs/design-system/styles/fonts/termina/Termina-Heavy.otf
