+++ libs/design-system/styles/fonts/termina/Termina-Light.otf
