+++ libs/design-system/styles/fonts/termina/Termina-Medium.otf
