+++ libs/design-system/styles/fonts/termina/Termina-Regular.otf
