+++ libs/design-system/styles/fonts/termina/Termina-Thin.otf
